moved online renderer to the scenes folder...separated dirlight and spotlight into separate files
