+++ src/app_includes.h
@@ -1,0 +1,3 @@
+#pragma once
+
+#define SHOW_LIGHTS 1
