+++ src/lights/light_probe.cpp
