+++ src/lights/light_probe.h
