+++ src/scene/online_renderer.h
@@ -19,4 +19,3 @@ struct online_renderer_t {
 void update_online_vertices(framebuffer_t& final_offline_fb);
 void init_online_renderer();
 void render_online(tex_id_t final_att, int render_depth = 0);
-// void render_online(framebuffer_t& final_offline_fb);
